Addresses issues with @PushStateNavigation if application is not deployed in the root context.
When using CDI or Spring addon navigation to views removes the application context from the URL. As a result @PushStateNavigation can not be used by many users whom cannot deploy their applications to non root contexts.

This effects a great number of users as most large sites root context are not available. For instance many *.gov domains have hundreds of applications running under sub contexts ie. *.gov/app1, *.gov/app2 etc.

As a workaround a user could implement a reverse proxy along with some response filtering rules but this is not desireable in many cases.

This has not been tested outside of the Spring addon.